photoapp/imageapp/serializers.py
from rest_framework import serializers
from .models import Photo


class PhotoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Photo
        fields = ['id', 'photo', 'qr_code', 'download_url', 'upload_time']
