photoapp/requirements.txt
amqp==5.2.0
asgiref==3.5.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.0.6
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
exceptiongroup==1.2.1
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.5
packaging==23.2
pillow==10.4.0
pluggy==1.5.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==8.0
redis==5.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
